--- src/browser-storage.ts
@@ -1,22 +1,0 @@
-import { IInitialState } from "./store/transactions";
-
-const KEY = "redux";
-
-export function loadState() {
-  try {
-    const serializedState = localStorage.getItem(KEY);
-    if (!serializedState) return undefined;
-    return JSON.parse(serializedState) as IInitialState;
-  } catch (e) {
-    return undefined;
-  }
-}
-
-export async function saveState(state: any) {
-  try {
-    const serializedState = JSON.stringify(state);
-    localStorage.setItem(KEY, serializedState);
-  } catch (e) {
-    // Ignore
-  }
-}
